requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
jinja2==3.1.4
asyncpg==0.29.0
python-dotenv==1.0.1
httpx==0.27.2
openai==1.43.0
tenacity==9.0.0
